test: move UserCreatorService to a correct folder
